atom and bond ids from u64 to u32: 20% memory saving in AtomicStructure

+++ lib/src/rust/api/structure_designer/structure_designer_api_types.dart
@@ -222,7 +222,7 @@ class APIEditAtomData {
   final APIEditAtomTool activeTool;
   final bool canUndo;
   final bool canRedo;
-  final BigInt? bondToolLastAtomId;
+  final int? bondToolLastAtomId;
   final int? replacementAtomicNumber;
   final int? addAtomToolAtomicNumber;
   final bool hasSelectedAtoms;
